requirements.txt
Django==2.1.3
django-cors-headers==2.4.0
django-jet==1.0.8
djangorestframework==3.9.0
mysql==0.0.2
mysqlclient==1.3.13
pytz==2018.7
